+++ src/modules/shell/utils.sh
@@ -1,4 +1,4 @@
-# CLI utilities software task
+# CLI utilities task
 
 [[ -n "${_MOD_UTILS_LOADED:-}" ]] && return 0
 _MOD_UTILS_LOADED=1
@@ -12,6 +12,6 @@ utils::check() { apt::list_check "$_UTILS_LIST"; }
 utils::status() { apt::list_status "$_UTILS_LIST"; }
 
 utils::apply() {
-    apt::list_wizard "Software > Configure Utilities" \
+    apt::list_wizard "Shell Tools > Utilities" \
         "CLI utilities packages" "$_UTILS_LIST"
 }